requirements.txt
# Core data & ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
imbalanced-learn==0.11.0

# Gradient boosting
xgboost==2.0.0
lightgbm==4.0.0

# Deep learning
torch==2.0.1
torchvision==0.15.2

# API & deployment
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Streaming
kafka-python==2.0.2

# Visualization & dashboard
streamlit==1.28.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Explainability
shap==0.42.1
lime==0.2.0

# Utilities
joblib==1.3.1
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Jupyter & notebooks
jupyter==1.0.0
jupyterlab==4.0.6
ipython==8.15.0
ipywidgets==8.1.0
